High-Level and Low-Level Language Difference
Differences Between High-Level language and Low-Level language

The high-level language is programmer friendly. On the contrary, low-level language is machine amiable which means it is interpreted by machines easily.

Low-level languages take time to execute whereas high-level languages execute at a faster pace.

High-level languages are converted into machine specific language with the help of a compiler. As against, with low-level languages, only an assembler is needed for assembly language.

When it comes to memory consumption the low-level languages are highly efficient while high-level languages are not.
A high-level language is easily understandable whereas low-level language cannot be interpreted easily as it contains a set of long series of 0’s and 1’s.

Low-level languages cannot run over different machines as these are not- portable and machine independent. In contrast, the programs written in high-level languages are portable and machine independent.

Debugging and maintenance are easier and high-level languages as compared to low-level languages.